Initial commit of Get-VerbList
I'm lazy, and when I'm writing functions, I frequently need to check "is this a valid verb?" and get-verb is too 'noisy' and I just want a sorted list.  So, that's what this does.  If I care about what the category is, I added a switch to do that too.  This is another one that could probably use some improvement.